feat(docker): Replace YouTubeDL-Material with MeTube

+++ scripts/hard-reset-container.sh
@@ -16,8 +16,10 @@ function main() {
     docker stop "$CONTAINER_NAME" &> /dev/null
     docker rm "$CONTAINER_NAME" &> /dev/null
 
-    echo "Deleting '$CONTAINER_NAME' container state..."
-    sudo rm -rf "$DOCKER_BASE_DIRECTORY/$CONTAINER_NAME"
+    if [[ -d "$DOCKER_BASE_DIRECTORY/$CONTAINER_NAME" ]]; then
+      echo "Deleting '$CONTAINER_NAME' container state..."
+      sudo rm -rf "$DOCKER_BASE_DIRECTORY/$CONTAINER_NAME"
+    fi
   else
     echo "Container name '$CONTAINER_NAME' not running, current state: '$CONTAINER_STATE'"
     exit 1
